@@ -18,20 +18,20 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^8.8.1
-  carousel_slider: ^2.2.1
-  url_launcher: ^5.5.2
-  scrollable_positioned_list: ^0.1.7
+  font_awesome_flutter: ^10.7.0
+  carousel_slider: ^4.2.1
+  url_launcher: ^6.2.5
+  scrollable_positioned_list: ^0.3.8
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.6
 
 dev_dependencies:
   flutter_test:
