@@ -83,13 +83,13 @@ class Pro extends StatelessWidget {
                   padding: const EdgeInsets.only(top: 16.0),
                   child: SizedBox(
                     width: double.infinity,
-                    child: FlatButton(
-                      textColor: ArgonColors.white,
-                      color: ArgonColors.info,
+                    child: TextButton(
+                      // textColor: ArgonColors.white,
+                      // color: ArgonColors.info,
                       onPressed: _launchURL,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(4.0),
-                      ),
+                      // shape: RoundedRectangleBorder(
+                      //   borderRadius: BorderRadius.circular(4.0),
+                      // ),
                       child: Padding(
                           padding: EdgeInsets.only(
                               left: 16.0, right: 16.0, top: 12, bottom: 12),
