@@ -9,9 +9,9 @@ class CardSquare extends StatelessWidget {
       this.tap = defaultFunc});
 
   final String cta;
-  final String img;
-  final Function tap;
-  final String title;
+  final String? img;
+  final Function()? tap;
+  final String? title;
 
   static void defaultFunc() {
     print("the function works!");
@@ -39,7 +39,7 @@ class CardSquare extends StatelessWidget {
                                   topLeft: Radius.circular(6.0),
                                   topRight: Radius.circular(6.0)),
                               image: DecorationImage(
-                                image: NetworkImage(img),
+                                image: NetworkImage(img!),
                                 fit: BoxFit.cover,
                               )))),
                   Flexible(
@@ -51,7 +51,7 @@ class CardSquare extends StatelessWidget {
                           mainAxisAlignment: MainAxisAlignment.spaceBetween,
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: [
-                            Text(title,
+                            Text(title ?? "",
                                 style: TextStyle(
                                     color: ArgonColors.header, fontSize: 13)),
                             Text(cta,
